chore: adds toggle, separator, navbar comp and changes files types to jsx
