+++ app/components/styles/separator.module.css
@@ -1,0 +1,18 @@
+.seperator {
+  font-size: var(--fs-sec);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+
+  gap: 1em;
+}
+
+.seperator::after {
+  display: block;
+  content: "";
+
+  width: 100%;
+  height: 2px;
+
+  background-color: black;
+}
