+++ app/components/styles/toggle.module.css
@@ -1,0 +1,32 @@
+.toggle_outer {
+  --height-toggle: 20px;
+  display: block;
+
+  width: 45px;
+  height: var(--height-toggle);
+  border-radius: 15px;
+
+  background-color: var(--col-acc-tri);
+  display: flex;
+  align-items: center;
+
+  padding: 0.2em;
+
+  cursor: pointer;
+}
+
+.toggle_inner {
+  display: block;
+  width: calc(var(--height-toggle) - 5px);
+  height: calc(var(--height-toggle) - 5px);
+  background-color: aquamarine;
+  border-radius: 50%;
+}
+
+.left {
+  margin-left: auto;
+}
+
+.right {
+  margin-right: auto;
+}
